(WIP) rename '_all_touch_moves.py' to '_rest_of_touch_moves.py'
