+++ asynckivy/__init__.py
@@ -5,4 +5,4 @@ from ._sleep import *
 from ._event import *
 
 from ._animation import *
-from ._all_touch_moves import *
+from ._rest_of_touch_moves import *
